+++ tests/testthat/test-S3.R
@@ -43,6 +43,14 @@ test_that("can construct data frame subclass", {
   expect_s3_class(df, "data.frame")
 })
 
+test_that("inherits() works with S7_S3_class", {
+  skip_unless_r("> 4.3.0")
+
+  expect_true(inherits(factor("a"), class_factor))
+  expect_false(inherits(1, class_factor))
+  expect_true(inherits(Sys.Date(), new_S3_class("Date")))
+})
+
 # Basic tests of validators -----------------------------------------------
 
 test_that("catches invalid factors", {